fix(term): terminfo errors when stdout is not a tty
Avoid calling terminfo when stdout is not a tty.
To know if a tty is used, the `$TERM` environment variable is tested and
should be set to the terminfo's database name. The usual test is to use the
POSIX command `test` on the file descriptor 1 (`test -t 1`). But simply
reading `$TERM` exactly avoids the error: `tput` reads this env variable.