new(library): macro-functions to proxy makefile recipes and targets.
Implementation of two new amazing macros:
  - `m.in/proxy/recipe(command, recipe)`
  - `m.in/proxy/target(command, target-prefix?)`

`m.in/proxy/recipe(command, recipe)` allows transparent proxying of make
**recipes**. It proxies a recipe through a given proxy command.
It enables `.ONESHELL:` which changes the recipe execution to use one shell
instead of one shell per line of recipe. The effect is global and cannot be
local to a single rule and recipe. When it appears anywhere in the Makefile
then **all** recipe lines **for each** target will be provided to a single
invocation of the shell `$(SHELL)`. Cf. its documentation for more details.
It also allows recipe piping like in more modern tools based on nodejs.

`m.in/proxy/target(command, target-prefix?)` allows another kind of proxying.
It creates a new target which proxies make calls to a command throught its
command line. It results in a target `proxy-%` that matches targets like
`make proxy-all` resulting in `command make all`. This is an explicit way of
proxying, while `m.in/proxy/recipe()` tries to make it transparent and more
selective at the recipe level. In this case, you manually trigger the proxy.