new(toolchain): docker toolchain
Very simple docker toolchain with some common commands needed so far.
But yet very powerful combined with the new proxy macro-functions:
  - `m.in/toolchain/docker/dockerize/target()`
  - `m.in/toolchain/docker/dockerize/recipe()`
Based on `m.in/proxy/{target, recipe}()`, they allow to decouple the need for a
build tool (make) and the need for an build environment (a docker container).
Thus, a Makefile can be independently written without docker support and then
simply added by proxying a set of user-defined recipes or targets to go through
a `docker run` of `docker exec`.