new(library): support for phony targets
Add support for phony targets so that it behaves like `make_explicit()` but
without the assumption the target is an actual file to be generated.

It allows toolchains/users to provide custom targets such as `make test`
or `make doc` to drive sub-commands, but still make benefit from the framework
by creating a transaction with the target's name.