fix(toolchain/docker): dockerfile directory as build context
Pass the Dockerfile's directory to `docker build` as build context
and allow writing paths relative to the Dockerfile instead of the
project's root direction `$(m.in/root)`. This is a more natural
style: `ADD etc /etc` vs `ADD path/from/project/root/etc /etc`.
These files are generally in the same directory as the Dockerfile.