new(core): forced clean list
If your maker really needs to destroy its target with `make
clean`, the new file list `m.in/clean/force` is an unmanaged
(by Makefile.in) list of files that will **always** be removed
with `make clean`.

As soon as you use a `make_*()` macro from the library to create
a target, it is considered a final target and it is moved to the
`m.in/distclean` list (and removed from the clean list).

The first use of this list is the docker toolchain with the tracking
file of a docker image which is created with `make_explicit()` while
the maker tried to put the tracking file in the clean list. Given the
documented behaviour of the library regarding target cleaning rules,
it can now force this at `make clean` time.