Docker image builds tracking files
A hidden directory `.m.in` is now used to store tracker files of
tools generating outputs not traceable with a file.
Every time the result is generated, touch the tracking file to
update its timestamp so that make considers it as up-to-date.

The first usage is for docker builds which create docker images.